Core: Update to 10.0.7
Port From (https://github.com/TrinityCore/TrinityCore/commit/922f60fa25569db05a7717b6b55568f321a02f64)

+++ Source/Game/DataStorage/Structs/A_Records.cs
@@ -167,7 +167,7 @@ namespace Game.DataStorage
         public float BoxYaw;
         public sbyte ShapeType;
         public short ShapeID;
-        public short AreaTriggerActionSetID;
+        public int AreaTriggerActionSetID;
         public sbyte Flags;
     }
