+++ Source/Game/DataStorage/Structs/D_Records.cs
@@ -45,9 +45,9 @@ namespace Game.DataStorage
         public DifficultyFlags Flags;
         public byte ItemContext;
         public byte ToggleDifficultyID;
-        public ushort GroupSizeHealthCurveID;
-        public ushort GroupSizeDmgCurveID;
-        public ushort GroupSizeSpellPointsCurveID;
+        public uint GroupSizeHealthCurveID;
+        public uint GroupSizeDmgCurveID;
+        public uint GroupSizeSpellPointsCurveID;
     }
 
     public sealed class DungeonEncounterRecord
